Enable traditional threading as an option (#3149)
In a previous PR, stubs were added to enable traditional threading in
the UFS.
This PR takes the next step to execute the model with traditional
threading.
Specifically;
1. Do not use the `_esmf` `ufs_configure` file for traditional
threading. This sets the `globalResourceControl` to `false` (traditional
threading) ... this was done in the previous PR
2. when calculating components pet bounds do not multiply with number of
threads also in `ufs_configure` ... this PR
3. `export OMP_NUM_THREADS=$UFS_THREADS` ... this was done in the
previous PR

The default behaviour to use ESMF-managed threading is retained.
Traditional threading can be achieved by toggling `USE_ESMF_THREADING`
flag in `config.fcst`

Closes #3122 - as in allows the model to be configured to run w/
traditional threading. Testing on C768 S2SW was successful on Hercules
(This was previously reported to be failing)

The use of traditional threading will be inefficient for model
components that are not thread-enabled and supported e.g. MOM6 and
CICE6, but since they require smaller number of tasks compared to ATM or
WAV, the CPU 'waste' is not huge.